# book-network/src/main/java/com/soufiane/book/feedBack/FeedBackRequest.java
package com.soufiane.book.feedBack;

import jakarta.validation.constraints.*;

public record FeedBackRequest(
        @Positive(message = "200")
        @Min(value = 0, message = "201")
        @Max(value = 5, message = "202")
        Double note,
        @NotNull(message = "203")
        @NotEmpty(message = "203")
        @NotBlank(message = "203")
        String comment,
        @NotNull(message = "204")
        Integer bookId
) {
}